rename to task3, use locations
