+++ build.gradle.kts
@@ -24,6 +24,7 @@ dependencies {
     implementation("io.ktor:ktor-server-netty:$ktorVersion")
     implementation("io.ktor:ktor-client-apache:$ktorVersion")
     implementation("io.ktor:ktor-client-jackson:$ktorVersion")
+    implementation("io.ktor:ktor-locations:$ktorVersion")
     implementation("com.h2database:h2:1.4.200")
     implementation("com.zaxxer:HikariCP:3.2.0")
     implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.9.7")
